refactor(patrol_cli): rename video_recording_manager.dart to android_video_recording_manager.dart
Match the file name to the AndroidVideoRecordingManager class and mirror
the iOS ios_video_recording_manager.dart naming.
