+++ packages/patrol_cli/lib/src/android/android_video_recording_manager.dart
